Fix null value support (#60)
* INT-532: Permit null input/output variables.

Updates input/output validation to permit values to be null.
This is done by overriding the functionality of the Draft4 Validator.

* Revert "INT-532: Permit null input/output variables."

This reverts commit 1fccee023cd7ecb1cc72723efa921e6541da1c3e.

python_jsonschema_objects should do this for us.

* Use a forked version of python_jsonschema_objects

This commit makes the sdk use a forked version of
python_jsonschema_objects.

Once my changes in that libary are accepted, the fork can be removed.

Strict schema validation is also enabled. Strict schema validation means
the required properties must be included, which is the behavior we want.

(This was previously working because of a side effect in the old
python_jsonschema_objects libary, where it would error. Now it no longer
errors, so we have to check that we want strict validation)

* Fix sample, Fix Docker build

* Use a branch for python-jsonschema-objects

* Pin certifi

* fix missing =